refactor: :recycle: common module 생성, 파일 구조 변경
- #15
