+++ app/src/common/models/common.coalition.model.ts
@@ -1,0 +1,21 @@
+import { Field, ID, ObjectType, registerEnumType } from '@nestjs/graphql';
+
+export enum CoaliltionName {
+  GUN,
+  GON,
+  GAM,
+  LEE,
+}
+
+registerEnumType(CoaliltionName, {
+  name: 'CoaliltionName',
+});
+
+@ObjectType()
+export class Coalition {
+  @Field((_type) => ID)
+  id: string;
+
+  @Field()
+  name: CoaliltionName;
+}
