--- app/src/home/models/examInfo.type.ts
@@ -1,13 +1,0 @@
-import { Field, Int, ObjectType } from '@nestjs/graphql';
-
-@ObjectType()
-export class ExamResult {
-  @Field((_type) => Int)
-  rank: number;
-
-  @Field((_type) => Int)
-  passCnt: number;
-
-  @Field((_type) => Int)
-  total: number;
-}
