+++ app/src/personalGeneral/personal.general.module.ts
@@ -1,6 +1,7 @@
 import { Module } from '@nestjs/common';
-import { PersonalGeneralResolver, UserProfileResolver } from './personal.general.resolver';
+import { PersonalGeneralResolver } from './personal.general.resolver';
 import { PersonalGeneralService } from './personal.general.service';
+import { UserProfileResolver } from './personal.general.userProfile.resolver';
 
 @Module({
   imports: [],
