+++ app/src/personalGeneral/personal.general.resolver.ts
@@ -1,6 +1,6 @@
 import { Int, Query, ResolveField, Resolver } from '@nestjs/graphql';
 import { EvalUserInfo, LevelGraph, LogtimeInfo, PersonalGeneral, TeamInfo } from './models/personal.general.model';
-import { UserProfile } from './models/personal.general.profile.model';
+import { UserProfile } from './models/personal.general.userProfile.model';
 import { PersonalGeneralService } from './personal.general.service';
 
 @Resolver((_of: unknown) => PersonalGeneral)
@@ -28,23 +28,13 @@ export class PersonalGeneralResolver {
     return await this.personalGeneralService.getTeamInfoByUid('99947');
   }
 
-  @ResolveField('levelGraphs', (_returns) => LevelGraph)
+  @ResolveField('levelGraphs', (_returns) => [LevelGraph])
   async getLevelHistory() {
     return await this.personalGeneralService.getLevelHistroyByUid('99947');
   }
-}
 
-@Resolver((_of: unknown) => UserProfile)
-export class UserProfileResolver {
-  constructor(private personalGeneralService: PersonalGeneralService) {}
-
-  @ResolveField('personalProfile', (_retuns) => UserProfile)
+  @ResolveField('userProfile', (_returns) => UserProfile)
   async getUserInfo() {
     return await this.personalGeneralService.getUserInfo('99947');
-  }
-
-  @ResolveField('levelRank', (_returns) => Int)
-  async getLevelRank() {
-    return 24;
   }
 }
