+++ app/src/personalGeneral/personal.general.userProfile.resolver.ts
@@ -1,0 +1,12 @@
+import { Int, ResolveField, Resolver } from '@nestjs/graphql';
+import { UserProfile } from './models/personal.general.userProfile.model';
+
+@Resolver((_of: unknown) => UserProfile)
+export class UserProfileResolver {
+  constructor() {}
+
+  @ResolveField('levelRank', (_returns) => Int)
+  async getLevelRank() {
+    return 24;
+  }
+}
